feat: refactor Geoname and CacheItem interfaces for improved type definitions and organization
